Quick and dirty CSS support. Closes #37.
Setting Watir.prefer_css = true will make watir-webdriver attempt to use CSS when possible, and fall back to XPath otherwise. For IE, this should bring a performance improvement for version 7 or higher (which supports document.querySelector). Note that using attribute selectors are not supported in IE's "Quirks" mode. To use attribute selectors, add a !DOCTYPE directive that specifies a standard-based document to the SUT.

For watirspec, the CSS selector is used in 53% of cases when enabled. There's probably room for improvement to make this percentage higher. When running watirspec, set the PREFER_CSS environment variable to enable the CSS translation, and set SELECTOR_STATS to get a report of what WebDriver locators are being used.