Optimize regexp selectors when using xpath
Regexp selectors are highly inefficient as they require additional round
trip requests (to retrieve attribute values) for each element that
matches the simplified selector. This optimization limits the number of
requests needed by converting leading and trailing runs of literals in
each regular expression to to an xpath `contains` predicate. Note that
problematic regular expressions such as case insensitive patterns and
those containing '|' are not converted.